+++ Documentation/saws_visualization.py
@@ -1,0 +1,186 @@
+import pydeck
+from geojson import Feature, Point, FeatureCollection, Polygon
+from cssaw_central.Session import Session
+
+from sklearn import preprocessing
+import pandas as pd
+import numpy as np
+import streamlit as st
+
+# method that takes a date in as an argument and reformats it to remove the day
+def reformatDate(date):
+    month = date[0:date.index("-")]
+    year = date[-4:]
+    if len(month) == 1:
+        month = "0{}".format(month)
+    
+    return "{}{}".format(year, month)
+
+@st.cache
+def get_monthly_saws_data(startDate, endDate):
+    """ Returns a pandas df with monthly saws data
+        args:
+            startDate ---- a string in the YYYYMMDD format for the start of the desired data
+            endDate ---- a string in the YYYYMMDD format for the end of the desired data
+            sess ---- a cssaw-central session used to query the database
+            station ---- a string representing the desired station. ex: 'A2H056'
+    """
+    startMonth, startDay, startYear = startDate[4:6], startDate[-2:], startDate[0:4]
+    endMonth, endDay, endYear = endDate[4:6], endDate[-2:], endDate[0:4]
+
+    # remove leading 0s from month and day entries to match the dates in the database
+    if startDay[0] == '0':
+        startDay = startDay[1:]
+    if startMonth[0] == '0':
+        startMonth = startMonth[1:]
+    if endDay[0] == '0':
+        endDay = endDay[1:]
+    if endMonth[0] == '0':
+        endMonth = endMonth[1:]
+
+    # parse the current month and year to be the starting month and year as integers
+    currentMonth = int(startMonth)
+    currentYear = int(startYear)
+
+    # get the dataframe for the starting month and year
+    currentDate = "{}-1-{}".format(currentMonth, currentYear)
+    query = "SELECT * from CENTRAL.saws_precipitation" + " WHERE `Date` LIKE '" + currentDate + "'"
+    # get data by date from the database through a query and set dataframe = to the results
+    dataFrame = sess.execute_query(query, pandas=True)
+
+    # increment currentMonth and make sure it stays in range of 12 months, incrementing the year after 12 months
+    currentMonth += 1
+    if currentMonth > 12:
+        currentMonth = 1
+        currentYear += 1
+
+    # make sure that the start date and end date are not within the same month and year
+    if startMonth != endMonth or startYear != endYear:
+        # loop through all the months and years until you reach the end month and year
+        while currentMonth != int(endMonth) and currentYear != endYear:
+            currentDate = "{}-1-{}".format(currentMonth, currentYear)
+            query = "SELECT * from CENTRAL.saws_precipitation" + " WHERE `Date` LIKE '" + currentDate + "'"
+
+            # increment currentMonth and make sure it stays in range of 12 months, incrementing the year after 12 months
+            currentMonth += 1
+            if currentMonth > 12:
+                currentMonth = 1
+                currentYear += 1
+
+            # get data by date from the database through a query and append results to the dataframe
+            dataFrame = dataFrame.append(sess.execute_query(query, pandas=True), ignore_index=True)
+
+        # handle the end month, year here
+        currentDate = "{}-1-{}".format(currentMonth, currentYear)
+        query = "SELECT * from CENTRAL.saws_precipitation" + " WHERE `Date` LIKE '" + currentDate + "'"
+        # get data by date from the database and append to the dataframe
+        dataFrame = dataFrame.append(pd.DataFrame(sess.execute_query(query, pandas=True)), ignore_index=True)
+
+    # The line below gets rid of the Day portion of the Date column
+    dataFrame["Date"] = dataFrame["Date"].apply(reformatDate) 
+
+    return dataFrame
+
+# Some c
+def create_geo_json(date, sess):
+    dataFrame = get_monthly_saws_data(date, date)
+    sess.conn.close()
+
+
+    # normalized_data = dataFrame["Rainfall (mm)"]
+    # # print(normalized_df.head())
+
+    # # min_max_scaler = preprocessing.MinMaxScaler()
+    # # normalized_data = min_max_scaler.fit_transform(normalized_data.reshape())
+    # print(pd.DataFrame(dataFrame["Rainfall (mm)"]).max() - pd.DataFrame(dataFrame["Rainfall (mm)"]).min() )
+    # dataFrame["normalizedCor"] = pd.DataFrame(dataFrame["Rainfall (mm)"]).apply(lambda x: (x-x.min())/(x.max()-x.min()))
+
+    print(dataFrame.head())
+    # normalized_df=(normalized_df-normalized_df.min())/(normalized_df.max()-normalized_df.min())
+    polygonSize = 0.02/2
+    geoJsonList = []
+    # print(normalized_df.head())
+    for index,row in dataFrame.iterrows():
+        # feature = Feature(geometry=Point((row['long'], row['Latitude'])), properties={"elevation":row["Rainfall (mm)"]})
+        feature = Feature(geometry=Polygon([[[row['Longitude']-polygonSize, row['Latitude']-polygonSize],[row['Longitude']-polygonSize, row['Latitude']+polygonSize],[row['Longitude']+polygonSize, row['Latitude']+polygonSize],[row['Longitude']+polygonSize, row['Latitude']-polygonSize]]]), 
+        properties={"elevation": row["Rainfall (mm)"], "normalizedElevation": row["Rainfall (mm)"]
+        })
+        geoJsonList.append(feature)
+
+    featureCollection = FeatureCollection(geoJsonList)
+    return featureCollection
+
+if __name__ == "__main__":
+    credentials = open('../credentials.txt', 'r')
+    username = credentials.readline().replace('\n','')
+    password = credentials.readline().replace('\n','')
+    host = credentials.readline().replace('\n','')
+    mapbox_api_key = credentials.readline().replace('\n','')
+    credentials.close()
+    
+    sess = Session(username,password, host, db='CENTRAL')
+    data = create_geo_json(20120101,sess)
+    
+    INITIAL_VIEW_STATE = pydeck.ViewState(Latitudeitude=-24.654950, Longitudeitude=29.3906515, zoom=7, max_zoom=16, pitch=45, bearing=0)
+
+    geojson = pydeck.Layer(
+        "GeoJsonLayer",
+        data = data,
+        opacity=0.8,
+        stroked=False,
+        filled=True,
+        extruded=True,
+        wireframe=True,
+        get_elevation="elevation*1500000",
+        get_radius=1000, 
+        get_fill_color="[255 ,100, normalizedElevation * 255, 255]", 
+        get_line_color=[255, 255, 255],
+    )
+
+
+    r = pydeck.Deck(layers=[ geojson], initial_view_state=INITIAL_VIEW_STATE)
+
+    r.to_html("saws_geojson_layer.html")
+
+
+
+# if __name__ == "__main__":
+#     st.title('Data Visualization')
+
+#     st.write("Use the side bar to select different dates")
+
+#     months = {
+#         "January": "01",
+#         "Febrauary": "02",
+#         "March": "03",
+#         "April": "04",
+#         "May": "05",
+#         "June": "06",
+#         "July": "07",
+#         "August": "08",
+#         "September": "09",
+#         "October": "10",
+#         "November": "11",
+#         "December": "12"
+#     }
+#     years = ["2012", "2013", "2014", "2015", "2016", "2017", "2018", "2019", "2020"]
+
+#     month = st.sidebar.selectbox("Pick a month", list(months.keys()))
+#     year = st.sidebar.selectbox("Pick a year:", years)
+
+#     "You selected: ", month, year
+
+#     startDate = "{}{}01".format(year, months[month])
+#     # use 1 month for testing
+#     endDate = startDate
+
+#     # use getter method to get data from database for selected date
+#     df = get_monthly_saws_data(startDate,endDate)
+    
+#     # df = pd.DataFrame({
+#     # 'Date': ['3-1-2012', '3-1-2012', '3-1-2012', '3-1-2012'],
+#     # 'Latitudeitude': [25.32, 25.36, 25.32, 25.36],
+#     # 'Longitude': [27.3, 27.3, 28.1, 28.1],
+#     # 'Precipitation': [0.002, 0.32, 0.091, 0.672]
+#     # })
+#     st.write(df.head())
